requirements.txt
Flask==2.3.2
Flask-Caching==2.3.0
python-dotenv==1.0.0